dags/requirements.txt
# See https://cloud.google.com/composer/docs/composer-versions#images-composer-3
# See https://cloud.google.com/composer/docs/versions-packages#composer-3-airflow-2-10-5-build-12
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-google==17.1.0
apache-airflow-providers-standard==1.5.0
